% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_microbiome.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Filter Taxa Based on Taxonomy Table Attributes}
\usage{
filter_taxa(tb, ...)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}

\item{...}{Logical predicates passed to \link[dplyr:filter]{dplyr::filter} on the taxonomy table.}
}
\value{
A filtered \code{tidy_microbiome} object with synchronized assays, taxonomy, and tree.
}
\description{
Filter Taxa Based on Taxonomy Table Attributes
}
