% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{from_s7}
\alias{from_s7}
\title{Convert S7 Formal Object to tidy_microbiome S3 Tibble}
\usage{
from_s7(s7_obj)
}
\arguments{
\item{s7_obj}{An object created by \code{\link[=to_s7]{to_s7()}}.}
}
\value{
A \code{tidy_microbiome} S3 object.
}
\description{
Coerces an S7 \code{tidy_microbiome_s7} object back into a standard S3 \code{tidy_microbiome} tibble.
}
\examples{
if (requireNamespace("S7", quietly = TRUE)) {
  data(gut_microbiome)
  tb_s7 <- to_s7(gut_microbiome)
  tb_restored <- from_s7(tb_s7)
  class(tb_restored)
}
}
