% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{from_vegan}
\alias{from_vegan}
\title{Construct tidy_microbiome from vegan Community Format}
\usage{
from_vegan(comm, env = NULL, tax_table = NULL, phy_tree = NULL)
}
\arguments{
\item{comm}{Community matrix or data frame with samples in rows and taxa in columns.}

\item{env}{Optional sample metadata data frame.}

\item{tax_table}{Optional taxonomy data frame.}

\item{phy_tree}{Optional phylogenetic tree.}
}
\value{
A \code{tidy_microbiome} object.
}
\description{
Construct tidy_microbiome from vegan Community Format
}
