% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{get_distance}
\alias{get_distance}
\title{Extract Distance Matrix from tidy_microbiome}
\usage{
get_distance(tb, metric = NULL)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}

\item{metric}{Optional character string naming the distance metric to extract.
If \code{NULL}, returns the most recently calculated distance matrix.}
}
\value{
An object of class \code{dist}.
}
\description{
Extract Distance Matrix from tidy_microbiome
}
