% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{get_tree}
\alias{get_tree}
\title{Get Phylogenetic Tree from tidy_microbiome}
\usage{
get_tree(tb)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}
}
\value{
An \code{ape::phylo} object, or \code{NULL} if none is attached.
}
\description{
Get Phylogenetic Tree from tidy_microbiome
}
