% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_core}
\alias{plot_core}
\title{Plot Core Microbiome Landscape}
\usage{
plot_core(core_res)
}
\arguments{
\item{core_res}{A data frame or tibble produced by \code{\link[=calc_core_microbiome]{calc_core_microbiome()}}.}
}
\value{
A \code{ggplot2::ggplot} object.
}
\description{
Visualizes the core microbiome distribution across prevalence and abundance.
}
