% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dbrda}
\alias{plot_dbrda}
\title{Plot Distance-Based Redundancy Analysis (db-RDA) Biplot}
\usage{
plot_dbrda(dbrda_res, color = NULL, shape = NULL, palette = "tidybiome")
}
\arguments{
\item{dbrda_res}{An object produced by \code{\link[=run_dbrda]{run_dbrda()}}.}

\item{color}{Optional metadata variable to color sample points.}

\item{shape}{Optional metadata variable for point shapes.}

\item{palette}{Palette name: \code{"tidybiome"} or \code{"nature"}.}
}
\value{
A \code{ggplot2::ggplot} object.
}
\description{
Plot Distance-Based Redundancy Analysis (db-RDA) Biplot
}
