% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{set_tree}
\alias{set_tree}
\title{Attach and Validate Phylogenetic Tree to tidy_microbiome}
\usage{
set_tree(tb, tree, prune = TRUE)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}

\item{tree}{An \code{ape::phylo} object.}

\item{prune}{Logical; if \code{TRUE}, automatically prunes the tree to match taxon IDs present in \code{tb}. Defaults to \code{TRUE}.}
}
\value{
A \code{tidy_microbiome} object with the attached tree.
}
\description{
Attach and Validate Phylogenetic Tree to tidy_microbiome
}
