% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{tax_table}
\alias{tax_table}
\title{Extract Taxonomy Table from tidy_microbiome}
\usage{
tax_table(x, ...)
}
\arguments{
\item{x}{A \code{tidy_microbiome} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \link[tibble:tbl_df]{tibble::tbl_df} with taxonomic lineage annotations.
}
\description{
Extract Taxonomy Table from tidy_microbiome
}
