% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tidybiome}
\alias{theme_tidybiome}
\title{Modern Minimalist Publication Theme for tidybiome}
\usage{
theme_tidybiome(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size (default: 11).}

\item{base_family}{Base font family.}
}
\value{
A \code{ggplot2} theme object.
}
\description{
Modern Minimalist Publication Theme for tidybiome
}
