% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{tidy_taxa}
\alias{tidy_taxa}
\title{Extract Taxonomy Annotations as a Tidy Tibble}
\usage{
tidy_taxa(tb)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}
}
\value{
A \link[tibble:tbl_df]{tibble::tbl_df} with \code{taxon_id} and taxonomic ranks.
}
\description{
Extract Taxonomy Annotations as a Tidy Tibble
}
