% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybiome-package.R
\docType{package}
\name{tidybiome-package}
\alias{tidybiome-package}
\alias{tidybiome}
\title{tidybiome: Tidyverse-Native Modern Microbiome Analysis Suite}
\description{
A modern, tidyverse-native Swiss Army knife for downstream microbiome analysis.
Implements state-of-the-art bioRxiv and literature methodologies for compositional
transformation (Robust CLR), alpha diversity (unified Hill numbers profile),
beta diversity (Robust Aitchison, Tree-Wasserstein, UniFrac), ordination (RPCA, cPCA),
multi-engine consensus differential abundance testing, constrained ordination (db-RDA),
and microbial co-occurrence networks.
}
\author{
\strong{Maintainer}: Tidybiome Authors \email{info@tidybiome.org}

Authors:
\itemize{
  \item Tidybiome Authors \email{info@tidybiome.org}
}

}
\keyword{internal}
