% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{to_phyloseq}
\alias{to_phyloseq}
\title{Convert tidy_microbiome to phyloseq Object}
\usage{
to_phyloseq(tb)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}
}
\value{
A \code{phyloseq::phyloseq} object.
}
\description{
Convert tidy_microbiome to phyloseq Object
}
