% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{to_tse}
\alias{to_tse}
\title{Convert tidy_microbiome to TreeSummarizedExperiment Object}
\usage{
to_tse(tb)
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}
}
\value{
A \code{TreeSummarizedExperiment::TreeSummarizedExperiment} object.
}
\description{
Convert tidy_microbiome to TreeSummarizedExperiment Object
}
