% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\name{to_vegan}
\alias{to_vegan}
\title{Convert tidy_microbiome to vegan Community and Environmental Format}
\usage{
to_vegan(tb, assay = "counts")
}
\arguments{
\item{tb}{A \code{tidy_microbiome} object.}

\item{assay}{Name of assay to extract. Defaults to \code{"counts"}.}
}
\value{
An S3 object of class \code{tidybiome_vegan} containing:
\itemize{
\item \code{comm}: Community matrix with samples in rows and taxa in columns.
\item \code{env}: Sample metadata tibble.
}
}
\description{
Convert tidy_microbiome to vegan Community and Environmental Format
}
