% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_microbiome.R
\name{validate_tidy_microbiome}
\alias{validate_tidy_microbiome}
\title{Validate Integrity of a tidy_microbiome Object}
\usage{
validate_tidy_microbiome(tb, verbose = FALSE)
}
\arguments{
\item{tb}{An object to validate.}

\item{verbose}{Logical. If \code{TRUE}, prints a confirmation message upon successful validation.
Defaults to \code{FALSE}.}
}
\value{
\code{TRUE} if valid, otherwise throws an error.
}
\description{
Validate Integrity of a tidy_microbiome Object
}
