% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_microbiome.R
\name{validate_tidy_microbiome}
\alias{validate_tidy_microbiome}
\title{Validate Integrity of a tidy_microbiome Object}
\usage{
validate_tidy_microbiome(tb)
}
\arguments{
\item{tb}{An object to validate.}
}
\value{
\code{TRUE} if valid, otherwise throws an error.
}
\description{
Validate Integrity of a tidy_microbiome Object
}
